[refactore] S3 Service(#48)
